+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroGradleSubplugin.kt
@@ -10,6 +10,7 @@ import org.jetbrains.kotlin.gradle.plugin.KotlinCompilerPluginSupportPlugin
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
 import org.jetbrains.kotlin.gradle.plugin.SubpluginArtifact
 import org.jetbrains.kotlin.gradle.plugin.SubpluginOption
+import org.jetbrains.kotlin.gradle.plugin.getKotlinPluginVersion
 
 public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
 
@@ -22,7 +23,31 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
   override fun getPluginArtifact(): SubpluginArtifact =
       SubpluginArtifact(groupId = "dev.zacsweers.metro", artifactId = "compiler", version = VERSION)
 
-  override fun isApplicable(kotlinCompilation: KotlinCompilation<*>): Boolean = true
+  override fun isApplicable(kotlinCompilation: KotlinCompilation<*>): Boolean {
+    val project = kotlinCompilation.target.project
+
+    // Check version and show warning by default.
+    val checkVersions =
+        project.extensions
+            .getByType(MetroPluginExtension::class.java)
+            .enableKotlinVersionCompatibilityChecks
+            .getOrElse(true)
+    if (checkVersions) {
+      val metroVersion = VersionNumber.parse(BASE_KOTLIN_VERSION)
+      val kotlinVersion = VersionNumber.parse(project.getKotlinPluginVersion())
+      if (metroVersion < kotlinVersion) {
+        project.logger.warn(
+            "Metro '$VERSION' is too old for Kotlin '$kotlinVersion'. " +
+                "Please upgrade Metro or downgrade Kotlin to '$BASE_KOTLIN_VERSION'.")
+      } else if (metroVersion > kotlinVersion) {
+        project.logger.warn(
+            "Metro '$VERSION' is too new for Kotlin '$kotlinVersion'. " +
+                "Please upgrade Kotlin to '$BASE_KOTLIN_VERSION'.")
+      }
+    }
+
+    return true
+  }
 
   override fun applyToCompilation(
       kotlinCompilation: KotlinCompilation<*>
@@ -35,10 +60,7 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         "dev.zacsweers.metro:runtime:$VERSION",
     )
     if (kotlinCompilation.implementationConfigurationName == "metadataCompilationImplementation") {
-      project.dependencies.add(
-          "commonMainImplementation",
-          "dev.zacsweers.metro:runtime:$VERSION",
-      )
+      project.dependencies.add("commonMainImplementation", "dev.zacsweers.metro:runtime:$VERSION")
     }
 
     val isJvmTarget =
@@ -60,7 +82,9 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         add(lazyOption("generate-hint-properties", extension.generateHintProperties))
         add(
             lazyOption(
-                "enable-top-level-function-injection", extension.enableTopLevelFunctionInjection))
+                "enable-top-level-function-injection",
+                extension.enableTopLevelFunctionInjection,
+            ))
         extension.reportsDestination.orNull
             ?.let { FilesSubpluginOption("reports-destination", listOf(it.asFile)) }
             ?.let(::add)
@@ -69,7 +93,8 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
           add(
               SubpluginOption(
                   "enable-dagger-runtime-interop",
-                  extension.interop.enableDaggerRuntimeInterop.getOrElse(false).toString()))
+                  extension.interop.enableDaggerRuntimeInterop.getOrElse(false).toString(),
+              ))
         }
 
         with(extension.interop) {